# Add sampling memory profiler mode