--- a/include/seastar/util/sampler.hh
+++ b/include/seastar/util/sampler.hh
@@ -14,82 +14,148 @@
  * limitations under the License.
  */
 
-#ifndef SRC_PROFILING_MEMORY_SAMPLER_H_
-#define SRC_PROFILING_MEMORY_SAMPLER_H_
+// This file has been originally been imported from:
+// https://cs.android.com/android/platform/superproject/+/013901367630d3ec71c9f2bb3f3077bd11585301:external/perfetto/src/profiling/memory/sampler.h
 
-#include <stdint.h>
+//
+// The code has been modified as follows:
+//
+//  - Integrated into seastar and adapted to coding style
+//  - Right now we don't account for samples multiple times (in case we have
+//  multiple loops of drawing from the exp distribution). The reason is that
+//  in our memory sampler we would have to store the weight in addition to the
+//  alloation site ptr as on free we need to know how much a sample accounted
+//  for. Hence, for now we simply always use the sampling interval.
+//  - Sampler can be turned "off" with a 0 sampling rate
+//  - The fast path is more optimized (as a consequence of the first point)
+//  - Provide a way to temporarily pause sampling
+//
+// Changes Copyright (C) 2023 ScyllaDB
 
-#include <atomic>
+#pragma once
+
 #include <random>
 
-#include "perfetto/ext/base/utils.h"
+// See also: https://perfetto.dev/docs/design-docs/heapprofd-sampling for more
+// background of how the sampler works
 
-namespace perfetto {
-namespace profiling {
-
-constexpr uint64_t kSamplerSeed = 1;
-
-std::default_random_engine& GetGlobalRandomEngineLocked();
-
-// Poisson sampler for memory allocations. We apply sampling individually to
-// each byte. The whole allocation gets accounted as often as the number of
-// sampled bytes it contains.
-//
-// The algorithm is inspired by the Chromium sampling algorithm at
-// https://cs.chromium.org/search/?q=f:cc+symbol:AllocatorShimLogAlloc+package:%5Echromium$&type=cs
-// Googlers: see go/chrome-shp for more details.
-//
-// NB: not thread-safe, requires external synchronization.
-class Sampler {
- public:
-  void SetSamplingInterval(uint64_t sampling_interval) {
-    sampling_interval_ = sampling_interval;
-    sampling_rate_ = 1.0 / static_cast<double>(sampling_interval_);
-    interval_to_next_sample_ = NextSampleInterval();
-  }
-
-  // Returns number of bytes that should be be attributed to the sample.
-  // If returned size is 0, the allocation should not be sampled.
-  //
-  // Due to how the poission sampling works, some samples should be accounted
-  // multiple times.
-  size_t SampleSize(size_t alloc_sz) {
-    if (PERFETTO_UNLIKELY(alloc_sz >= sampling_interval_))
-      return alloc_sz;
-    return static_cast<size_t>(sampling_interval_ * NumberOfSamples(alloc_sz));
-  }
-
-  uint64_t sampling_interval() const { return sampling_interval_; }
-
- private:
-  int64_t NextSampleInterval() {
-    std::exponential_distribution<double> dist(sampling_rate_);
-    int64_t next = static_cast<int64_t>(dist(GetGlobalRandomEngineLocked()));
-    // We approximate the geometric distribution using an exponential
-    // distribution.
-    // We need to add 1 because that gives us the number of failures before
-    // the next success, while our interval includes the next success.
-    return next + 1;
-  }
-
-  // Returns number of times a sample should be accounted. Due to how the
-  // poission sampling works, some samples should be accounted multiple times.
-  size_t NumberOfSamples(size_t alloc_sz) {
-    interval_to_next_sample_ -= alloc_sz;
-    size_t num_samples = 0;
-    while (PERFETTO_UNLIKELY(interval_to_next_sample_ <= 0)) {
-      interval_to_next_sample_ += NextSampleInterval();
-      ++num_samples;
+class sampler {
+public:
+    sampler() : random_gen(rd_device()) {
+        set_sampling_interval(0);
     }
-    return num_samples;
-  }
+    /// Sets the sampling interval in bytes. Setting it to 0 means to never sample
+    void set_sampling_interval(uint64_t sampling_interval) {
+        sampling_interval_ = sampling_interval;
+        if (sampling_interval_ == 0) {
+            // Set the interval very large. This means in practice we will
+            // likely never get this below zero and hence it's unlikely we will
+            // ever have to run the reset path with sampling off
+            interval_to_next_sample_ = std::numeric_limits<int64_t>::max();
+            return;
+        }
+        sampling_rate_ = 1.0 / static_cast<double>(sampling_interval_);
+        interval_to_next_sample_ = next_sampling_interval();
+    }
+    /// Returns true if this allocation of size `alloc_size` should be sampled
+    bool should_sample(size_t alloc_size) {
+        // We need to check the sampling_interval as 0 means off. Also we need
+        // to check whether this allocation has brought the interval
+        // below/to 0. We chose to check the interval_to_next_sample first. If
+        // heap profiling is compiled in it's likely we are using it as well.
+        // Hence, lets do the interval decrement and check first and only check
+        // whether sampling is off after.
+        interval_to_next_sample_ -= alloc_size;
+        if (interval_to_next_sample_ > 0) {
+            return false;
+        }
+        reset_interval_to_next_sample(alloc_size);
+        return sampling_interval_ != 0;
+    }
 
-  uint64_t sampling_interval_;
-  double sampling_rate_;
-  int64_t interval_to_next_sample_;
+    uint64_t sampling_interval() const { return sampling_interval_; }
+
+    /// How much should an allocation of size `allocation_size` count for
+    size_t sample_size(size_t allocation_size) const {
+        return std::max(allocation_size, sampling_interval_);
+    }
+
+    /// RAII class to temporarily pause sampling
+    struct disable_sampling_temporarily {
+        disable_sampling_temporarily() = default;
+        disable_sampling_temporarily(sampler& sampler)
+        : sampler_(&sampler)
+        , previous_sampling_interval_(sampler_->sampling_interval_)
+        , previous_sampling_rate_(sampler_->sampling_rate_)
+        , previous_interval_to_next_sample_(sampler_->interval_to_next_sample_) {
+            sampler_->set_sampling_interval(0);
+        }
+
+        ~disable_sampling_temporarily() {
+            if (sampler_) {
+                sampler_->sampling_interval_ = previous_sampling_interval_;
+                sampler_->sampling_rate_ = previous_sampling_rate_;
+                sampler_->interval_to_next_sample_ = previous_interval_to_next_sample_;
+            }
+        }
+
+    private:
+        sampler* sampler_ = nullptr;
+        uint64_t previous_sampling_interval_ = 0; // sampling interval before pausing
+        double previous_sampling_rate_ = 0; // sampling rate before pausing
+        int64_t previous_interval_to_next_sample_ = 0; // interval to next sample before pausing
+    };
+
+    /// Pauses sampling temporarily until the returned object is destroyed. This
+    /// is more efficient and statisically more correct than doing a back and
+    /// fourth of set_sampling_interval(0) and set_sampling_interval(RATE). The
+    /// reason is that that would reset the interval to the next sample and
+    /// force a reevaluation of the exponential distribution. This method avoids
+    /// that.
+    disable_sampling_temporarily pause_sampling() {
+        return disable_sampling_temporarily(*this);
+    }
+
+private:
+    /// Resets interval_to_next_sample_ by repeatedly drawing from the
+    /// exponential distribution given an allocation of size `alloc_size`
+    /// breached the current interval
+    void reset_interval_to_next_sample(size_t alloc_size)
+    {
+        if (sampling_interval_ == 0) { // sampling is off
+            interval_to_next_sample_ = std::numeric_limits<int64_t>::max();
+        }
+        else {
+            // Large allocations we will just consider in whole. This avoids
+            // having to sample the distribution too many times if a large alloc
+            // took us very negative we just add the alloc size back on
+            if (alloc_size > sampling_interval_) {
+                interval_to_next_sample_ += alloc_size;
+            }
+            else {
+                while (interval_to_next_sample_ < 0) {
+                    interval_to_next_sample_ += next_sampling_interval();
+                }
+            }
+        }
+    }
+
+    int64_t next_sampling_interval() {
+        std::exponential_distribution<double> dist(sampling_rate_);
+        int64_t next = static_cast<int64_t>(dist(random_gen));
+        // We approximate the geometric distribution using an exponential
+        // distribution.
+        // We need to add 1 because that gives us the number of failures before
+        // the next success, while our interval includes the next success.
+        return next + 1;
+    }
+
+    uint64_t sampling_interval_; // Sample every N bytes ; 0 means off
+    double sampling_rate_; // 1 / sampling_interval_ ; used by the exp distribution
+    // How many bytes remain to be allocated before we take a sample.
+    // Specifically, if this member has value N, a sample will be taken of the allocation
+    // that allocates the Nth+1 byte.
+    int64_t interval_to_next_sample_;
+    std::random_device rd_device;
+    std::mt19937_64 random_gen;
 };
-
-}  // namespace profiling
-}  // namespace perfetto
-
-#endif  // SRC_PROFILING_MEMORY_SAMPLER_H_