This series changes the existing heap profiling to support a sampling mode, added
by @StephanDollberg and includes also some optimizations of the allocation fast paths.

The following description of the sampling profiler comes from Stephan:

Instead of tracking every allocation, it will track a sampled subset of allocations. This
is useful for profiling in production systems where the overhead of tracking every
allocation is too high.

A sample is taken at a set sample rate which feeds into an exponential
distribution. Hence we avoid falling into patterns that would potentially make
the profile biased otherwise. This approach is inspired by the android heapprofd
memory profiler.

By adjusting the sample rate it's possible to tune the overhead of the memory
profiler. The idea is to make this an always-on production profiler that can
give insight into allocations in OOM scenarios with negliable overhead.

Starting with the latest release of Redpanda we are using this in a default
always-on mode with a sampling period of 3MB. This is a rather conservative
setting and results in no measurable overhead while still giving good resolution
for the top allocation contributors.

We also add an API to retrieve the current heap profile programmatically. In
Redpanda we use this to print the top sites on OOM using the additional
diagnostics API. Further it allows easier retrieval of the data in a
dev/profiling fashion.

As today heap profiler can be compiled in or not (using `--heap-prof` at 
./configure.py time) and can also be enabled or disabled at runtime if it
has been compiled in.

The individual commits contain a more detailed description of the respective
changes. Note the heap profiler and all added codepaths still require explicit
opt in via the extra define.

This series also includes some optimizations that especially improve
small allocations and hence we end up faster than the current state even with
the profiler on, under the assumption that small allocations (which here means
<= 16 KiB) are considerably more frequent than larger allocations, which is 
overwhelmingly the case for most applications.

For example, a new/delete pair for a small allocation prior to this series take on
average ~104 instructions, versus ~73 after _with heap profiling on_ or ~67 with
it compiled out. Performance measured in time rather than instructions generally
follows the same improvement factors as time, though with more variance.

To aid with analysis of the above we also added a set of microbenchmarks. Below
you will find the selected key numbers from this benchmark comparing current
master and this patch set with the profiler compiled in and
out. 

Note profiler-on numbers are taken from a run with very high sample period
to avoid actually taking any samples. Sample creation is fully dominated by
backtrace creation so this avoids noise.

With the profiler enabled at compile time the runtime performance is identical 
between "compiler turned off at runtime" and "very large sample period such
that zero samples are actually taken". That is, there is no additional check for
sampler enablement on the fast path, only the single check to see if the sampling
period has been reached.

There are a lot of profiling results, so I'll share first this summary of a single 
test, which uses `new` to allocate many small pointers and then deletes them.

The configurations show are:

*Baseline:* Current `master` without any of the changes in this series.
*New-OFF:* This series but with heap profiling disabled at compile time.
*New-ON:* This series and with heap profiling enabled at compile time.

First the instruction counts:

| Config | Compiler | test | instructions |
| - | - | - | -: | 
| Baseline | gcc-11   | alloc_bench.op_new_delete                | 97.6 |
| New-OFF  | gcc-11   | alloc_bench.op_new_delete                | 74.6 |
| New-ON   | gcc-11   | alloc_bench.op_new_delete                | 82.6 |
| Baseline | clang-16 | alloc_bench.op_new_delete                | 104.6 |
| New-OFF  | clang-16 | alloc_bench.op_new_delete                | 67.5|
| New-ON   | clang-16 |alloc_bench.op_new_delete                |  73.5 |

Next, runtimes on Skylake and Zen 4, these are from separate runs than the instruction counts above as I found that enabling instruction count measurement disturbed (i.e., increased) the timings by a significant amount, probably because some part of the perf syscall cost ends up timed.

| Config | Arch | Compiler | test                                     |  iterations |      median |         mad |         min |         max |
|-|-|-| -                                        |           - |           -: |           -: |           -: |           -: |
| Baseline | Skylake | clang-16  | alloc_bench.op_new_delete                |   134081000 |     7.376ns |     0.005ns |     7.370ns |     7.534ns |
| New-OFF  | Skylake | clang-16  | alloc_bench.op_new_delete                |   146836000 |     6.586ns |     0.008ns |     6.577ns |     6.765ns |
| New-ON   | Skylake | clang-16  | alloc_bench.op_new_delete                |   193241000 |     5.113ns |     0.004ns |     5.110ns |     5.246ns |
||
| Baseline | Zen4 | gcc-11   | alloc_bench.op_new_delete                |   207025000 |     4.796ns |     0.000ns |     4.795ns |     4.797ns |
| New-OFF  | Zen4 | gcc-11   | alloc_bench.op_new_delete                |   305103000 |     3.241ns |     0.000ns |     3.240ns |     3.241ns |
| New-ON   | Zen4 | gcc-11   | alloc_bench.op_new_delete                |   302127000 |     3.271ns |     0.000ns |     3.271ns |     3.273ns |
||
| Baseline | Zen4 | clang-16 | alloc_bench.op_new_delete                |   212207000 |     4.663ns |     0.000ns |     4.661ns |     4.663ns |
| New-OFF  | Zen4 | clang-16 | alloc_bench.op_new_delete                |   358628000 |     2.742ns |     0.001ns |     2.740ns |     2.745ns |
| New-ON   | Zen4 | clang-16 | alloc_bench.op_new_delete                |   347325000 |     2.833ns |     0.003ns |     2.830ns |     2.836ns |

The primary anomalous result is New-OFF on clang-16 on Skylake (also shows up on Alder Lake, not shown). The runtime is considerably slower than the instruction count would suggest (i.e., it is lower IPC than other configurations) and I don't see any problem in a manual inspection of the assembly. The same function is fast on Zen 2 and Zen 4. I believe it is probably an Intel-specific alignment effect. I did move the bottom of the test loop around to avoid the jcc erratum (results not shown) which improved the results but the result was still slower than expected. In any case, even this configuration is faster than the baseline and I expect the issue to go away at some point.

Large allocations are a bit slower, very slightly when heap profiling is off at compile time and more when they are enabled, e.g., here with 32,000 byte allocations:

*Baseline*
```
test                                      iterations      median         mad         min         max      allocs       tasks        inst
alloc_bench.alloc_free_large                72444000    13.125ns     0.010ns    13.073ns    13.140ns       1.000       0.000       336.1
```

*New-OFF*
```
test                                      iterations      median         mad         min         max      allocs       tasks        inst
alloc_bench.alloc_free_large                71368000    13.334ns     0.004ns    13.287ns    13.338ns       1.000       0.000       343.0
```

*New-ON*
```
test                                      iterations      median         mad         min         max      allocs       tasks        inst
alloc_bench.alloc_free_large                65224000    14.623ns     0.002ns    14.604ns    14.628ns       1.000       0.000       399.5
```

Such allocations are rare compared to small allocations, however, and in any case the relative cost of allocation goes down compared to the work done on the allocated memory itself (e.g., writing into it) and that ratio becomes ~negligible except for an application which allocates large regions but then does not use them.

We have many micro-benchmarks but this description is already quite long: let me know what you think would be useful.