@avikivity 2023-10-30
Looks good. Please make newly introduced public APIs doxygen documented and verify the rendering.