@StephanDollberg 2023-10-31
> Looks good. Please make newly introduced public APIs doxygen documented and verify the rendering.

I have added more of the doxygen parameters and checked the generated html.