@travisdowns 2023-12-04
Force https://github.com/scylladb/seastar/commit/03e9966729c727d850795cc5facaa49f716cb1e3 was to update also the `sampled_memory_profile` stub method to the raw interface.