@tchaikov 2023-12-23
> @tchaikov and @michoecho I think the most recent round of feedback has been addressed. If you can let us know if there is anything outstanding that we haven't addressed it would be much appreciated!

@travisdowns hi Travis, i think we are pending on a maintainer (hey @avikivity !) to review and merge this change. since maintainers are consistently been preempted by other tasks, guess a visible latency is expected =(