@michoecho 2023-12-24
> I'm waiting for @michoecho's blessing

Ouch, I didn't think you would wait for me again. The first review was already an approval, after all. I only left some nitpicks, they didn't deserve to block anything in the first place.

I can't see the changes since the last review (since github already GCd the hash I reviewed...), but I see that my comments have been addressed, so — assuming that no regression has sneaked in in the meantime — I (still) approve.

Sorry for stalling this PR for so long. I have no excuse — I should keep better track of review requests.