@tchaikov 2023-12-30
@avikivity hi Avi, ping?