@travisdowns 2024-01-02
@avikivity your feedback is addressed.

I've also pulled into one very small edge-case fix we've applied downstream as: https://github.com/scylladb/seastar/pull/1904/commits/a3548bcbbfa3ec93c04e7d8a65646072c8a35173