@travisdowns 2024-01-02
A big thank you for all the reviews and collaboration here. Hope this feature proves useful to you: it has already paid off nicely in Redpanda (the key is to emit some stacks in the OOM error message, which is up to the application currently).